+++ skbio/util/__init__.py
@@ -44,6 +44,15 @@ Plotting utilities
    PlottableMixin
 
 
+Randomization
+-------------
+
+.. autosummary::
+   :toctree: generated/
+
+   get_rng
+
+
 Decorators
 ----------
 
@@ -76,7 +85,6 @@ Generally useful functionality that doesn't fit in more specific locations.
 .. autosummary::
    :toctree: generated/
 
-   get_rng
    cardinal_to_ordinal
    find_duplicates
    safe_md5
@@ -91,7 +99,8 @@ Generally useful functionality that doesn't fit in more specific locations.
 # The full license is in the file LICENSE.txt, distributed with this software.
 # ----------------------------------------------------------------------------
 
-from ._misc import cardinal_to_ordinal, find_duplicates, safe_md5, get_rng
+from ._misc import cardinal_to_ordinal, find_duplicates, safe_md5
+from ._random import get_rng
 from ._testing import (
     get_data_path,
     assert_ordination_results_equal,
