+++ skbio/util/_docstring.py
@@ -6,6 +6,8 @@
 # The full license is in the file LICENSE.txt, distributed with this software.
 # ----------------------------------------------------------------------------
 
+"""Docstring utilities."""
+
 import re
