+++ skbio/util/_warning.py
@@ -6,6 +6,8 @@
 # The full license is in the file LICENSE.txt, distributed with this software.
 # ----------------------------------------------------------------------------
 
+"""Warning utilities."""
+
 from warnings import warn, simplefilter
