+++ skbio/util/tests/test_misc.py
@@ -6,12 +6,9 @@
 # The full license is in the file LICENSE.txt, distributed with this software.
 # ----------------------------------------------------------------------------
 
-import io
-import unittest
+from unittest import TestCase, main
+from io import BytesIO
 import re
-
-import numpy as np
-import numpy.testing as npt
 
 from skbio.util._misc import (
     cardinal_to_ordinal,
@@ -20,11 +17,10 @@ from skbio.util._misc import (
     MiniRegistry,
     chunk_str,
     resolve_key,
-    get_rng,
 )
 
 
-class TestMiniRegistry(unittest.TestCase):
+class TestMiniRegistry(TestCase):
     def setUp(self):
         self.registry = MiniRegistry()
 
@@ -138,7 +134,7 @@ class TestMiniRegistry(unittest.TestCase):
         self.assertEqual(obs, exp)
 
 
-class ResolveKeyTests(unittest.TestCase):
+class ResolveKeyTests(TestCase):
     def test_callable(self):
         def func(x):
             return str(x)
@@ -163,7 +159,7 @@ class ResolveKeyTests(unittest.TestCase):
             resolve_key({'foo': 1}, 'foo')
 
 
-class ChunkStrTests(unittest.TestCase):
+class ChunkStrTests(TestCase):
     def test_even_split(self):
         self.assertEqual(chunk_str('abcdef', 6, ' '), 'abcdef')
         self.assertEqual(chunk_str('abcdef', 3, ' '), 'abc def')
@@ -190,18 +186,18 @@ class ChunkStrTests(unittest.TestCase):
             chunk_str('abcdef', -42, ' ')
 
 
-class SafeMD5Tests(unittest.TestCase):
+class SafeMD5Tests(TestCase):
     def test_safe_md5(self):
         exp = 'ab07acbb1e496801937adfa772424bf7'
 
-        fd = io.BytesIO(b'foo bar baz')
+        fd = BytesIO(b'foo bar baz')
         obs = safe_md5(fd)
         self.assertEqual(obs.hexdigest(), exp)
 
         fd.close()
 
 
-class CardinalToOrdinalTests(unittest.TestCase):
+class CardinalToOrdinalTests(TestCase):
     def test_valid_range(self):
         # taken and modified from http://stackoverflow.com/a/20007730/3776794
         exp = ['0th', '1st', '2nd', '3rd', '4th', '5th', '6th', '7th', '8th',
@@ -218,7 +214,7 @@ class CardinalToOrdinalTests(unittest.TestCase):
             cardinal_to_ordinal(-1)
 
 
-class TestFindDuplicates(unittest.TestCase):
+class TestFindDuplicates(TestCase):
     def test_empty_input(self):
         def empty_gen():
             yield from ()
@@ -248,83 +244,5 @@ class TestFindDuplicates(unittest.TestCase):
         self.assertEqual(find_duplicates(gen()), set(['a', 2]))
 
 
-class TestGetRng(unittest.TestCase):
-
-    def test_get_rng(self):
-
-        # returns random generator
-        res = get_rng()
-        self.assertIsInstance(res, np.random.Generator)
-
-        # seed is Python integer
-        res = get_rng(42)
-        obs = np.array([res.integers(100) for _ in range(5)])
-        exp = np.array([8, 77, 65, 43, 43])
-        npt.assert_array_equal(obs, exp)
-
-        # seed is NumPy integer
-        res = get_rng(np.uint8(42))
-        obs = np.array([res.integers(100) for _ in range(5)])
-        npt.assert_array_equal(obs, exp)
-
-        # seed is new Generator
-        res = get_rng(res)
-        obs = np.array([res.integers(100) for _ in range(5)])
-        exp = np.array([85, 8, 69, 20, 9])
-        npt.assert_array_equal(obs, exp)
-
-        # test if integer seeds are disjoint
-        obs = [get_rng(i).integers(1e6) for i in range(10)]
-        exp = [850624, 473188, 837575, 811504, 726442,
-               670790, 445045, 944904, 719549, 421547]
-        self.assertListEqual(obs, exp)
-
-        # no seed: use current random state
-        np.random.seed(42)
-        res = get_rng()
-        obs = np.array([res.integers(100) for _ in range(5)])
-        exp = np.array([90, 11, 93, 94, 34])
-        npt.assert_array_equal(obs, exp)
-
-        # reset random state to reproduce output
-        np.random.seed(42)
-        res = get_rng()
-        obs = np.array([res.integers(100) for _ in range(5)])
-        npt.assert_array_equal(obs, exp)
-
-        # call also advances current random state
-        np.random.seed(42)
-        self.assertEqual(np.random.randint(100), 51)
-        res = get_rng()
-        self.assertEqual(np.random.randint(100), 14)
-
-        # seed is legacy RandomState
-        res = get_rng(np.random.RandomState(42))
-        obs = np.array([res.integers(100) for _ in range(5)])
-        npt.assert_array_equal(obs, exp)
-
-        # test if legacy random states are disjoint
-        obs = [get_rng(np.random.RandomState(i)).integers(1e6) for i in range(5)]
-        exp = [368454, 346004, 189187, 324799, 924851]
-        self.assertListEqual(obs, exp)
-
-        # invalid seed
-        msg = 'Invalid seed. It must be an integer or a random generator instance.'
-        with self.assertRaises(ValueError) as cm:
-            get_rng('hello')
-        self.assertEqual(str(cm.exception), msg)
-
-        # mimic legacy NumPy
-        delattr(np.random, 'Generator')
-        msg = ('The installed NumPy version does not support random.Generator. '
-               'Please use NumPy >= 1.17.')
-        with self.assertRaises(ValueError) as cm:
-            get_rng()
-        self.assertEqual(str(cm.exception), msg)
-        with self.assertRaises(ValueError) as cm:
-            get_rng('hello')
-        self.assertEqual(str(cm.exception), msg)
-
-
 if __name__ == '__main__':
-    unittest.main()
+    main()
